+++ components/icon/icon.scss
@@ -1,0 +1,3 @@
+//Load fontawesome here
+
+@import "~@fortawesome/fontawesome-free/scss/fontawesome";
