--- components/modal/card.twig
@@ -1,4 +1,0 @@
-<div class="uk-card {{type}} {{hover ? 'uk-card-hover' : ''}} uk-card-body">
-    <h3 class="uk-card-title">{{title}}</h3>
-    {{body}}
-</div>
