+++ components/modal/modal.twig
@@ -1,0 +1,20 @@
+<div id="{{id}}" class="uk-flex-top" uk-modal>
+	<div class="uk-modal-dialog {{ width }} {{ vertical ? 'uk-margin-auto-vertical' : '' }}" uk-overflow-auto>
+		<button class="uk-modal-close-default" type="button" uk-close></button>
+		{% if header %}
+			<div class="uk-modal-header">
+		    	<h2 class="uk-modal-title">{{title}}</h2>
+		    </div>
+		{% else %}
+			<h2 class="uk-modal-title">{{title}}</h2>
+		{% endif %}
+    	<div class="uk-modal-body">
+		    {{body}}
+		</div>
+		{% if footer %}
+			<div class="uk-modal-footer {{footer-align}}">
+				{{footer.body}}
+			</div>
+		{% endif %}
+	</div>
+</div>
